Intraday
If too many requests are made to the API, the already existing data will be overwritten. Now the data will not be overwritten and after a random time (1 to 9 minutes) the data will be tried one more time. If no file exists it will be created with "null" values.

-> Intraday Data

Crypto and Shares